requirements.txt
flask
streamlit
requests 
openai 
python-dotenv 